Implement get_value method to config objects
Which should allow retrieving values of nested configs even if some of
the nested elements are not instances of ParsedConfig themselves, but
rather, for instance, dictionaries.

Since we are allowing arbitrary configs to be added, even if they don't
have a corresponding model in the config_parser.py file, and these
configs without models become dictionaries, it is important to have
a way to access the items of these arbitrary configs in a standard way.