Fix issue in tasks.base.Task.get_task_setting
That would cause it to fail with AtributeError if the name of the class
representing the task didn't match the name of the config sections
corresponding to that task. This is the case, for example, for the
BinaryTask, whose expected config section name seems to be
"tasks.binary" instead of "task.BinaryTask".

Moved function `_get_name` from file `discover_task.py` to `base.py` to
avoid circular imports.