Code Refactoring (#120)
* rename/move: config --> deode/data/config_files

* rename/move: nam --> deode/data/namelists

* move/rename: ecf --> deode/templates/ecflow

* Remove .pre-commit-config.yaml file

* mv: config_file_schemas->data/config_file_schemas

* mv stand_alone.py out of the ecflow directory

* Fix some docs referring to the reshuffled files

* Fix paths to moved files in lakeheaven config

* Fix paths to moved files in test_main.py

* Replace uses of `Exception` with specific errors

This fixes #118

* Prevent CI from running tests/checks on forks

To prevent these from being doube-executed when we submit a PR or
push a commit to an existing PR (these were running in the fork
and also in the main repo).

* Hide coverage file created when running tests

To reduce clutter in the repo.

* Minor fixes to README.md

* Portable fix for running CI only within DEODE-NWP

* Fix name of coverage file in github actions config

* Fix imports to package's own stuff

Better to use relative imports than to import from `deode`.

Lint the order of the imports as well (using isort).

* Pass codecov patch check if coverage >= 80%

Otherwise, it may be difficult to add/modify code when the repo has a
higher coverage overall. See
<https://github.com/DEODE-NWP/Deode-Prototype/pull/120#issuecomment-1461841419>

* Install package, not just flakeheaven, in CI lint

Use poetry to install the whole package when running the linting CI,
instead of just installing flakeheaven using pip. The reason is that, by
using pip, the installed version may differ from the one specified in
the package, thus leading to possible cases where linting passes locally
but not in the github actions.

To make sure this is not too slow, caching of poetry dependencies have
been enabled using `actions/setup-python@v4`, so that the same
environment used in the tests is used in the linting task too.

* Include black & isort in the linting workflow

This closes #26

* Fix "cannot have both the `uses` and `run` keys"

* Make poetry write lock file before installing

As actions/setup-python@v4 seems to need it.

* Install poetry with their recommended method

* Add poetry to the CI's PATH

* Yet another attempt to put poetry in the CI's PATH

* Install poetry using pipx

* Another attempt at fixing poetry tests CI configs

* Use p3.9 in github test workflow

To try to get rid of the following errors when using the installed
python:

python3: /lib/x86_64-linux-gnu/libc.so.6: version `GLIBC_2.34' not found (required by python3)
python3: /lib/x86_64-linux-gnu/libm.so.6: version `GLIBC_2.35' not found (required by /__t/Python/3.8.16/x64/lib/libpython3.8.so.1.0)
python3: /lib/x86_64-linux-gnu/libc.so.6: version `GLIBC_2.33' not found (required by /__t/Python/3.8.16/x64/lib/libpython3.8.so.1.0)
python3: /lib/x86_64-linux-gnu/libc.so.6: version `GLIBC_2.32' not found (required by /__t/Python/3.8.16/x64/lib/libpython3.8.so.1.0)
python3: /lib/x86_64-linux-gnu/libc.so.6: version `GLIBC_2.34' not found (required by /__t/Python/3.8.16/x64/lib/libpython3.8.so.1.0)

* Revert to py3.8. Remove python setup step in CI

* Attempt at setting up workflow dependencies

* Activate local venv before running pytest

* Run pytest using poetry run

Trying `poetry shell` leads to the following error:

"(25 'inappropriate ioctl for device')"

* Run linting commands with poetry

* Add back poetry configs to linting workflow

* Organise deps in groups in pyproject.toml

As described in
<https://python-poetry.org/docs/master/managing-dependencies/#dependency-groups>

* Use a main workflow to manage all others

* Make sure flakeheaven cache is empty

* Pass codecov patch check if coverage >= 80%

This fixes a previous commit where the keyword "project" was used
instead of "patch" in the codecov.yml file.

* Move codecov token specs to codecov job in CI

* Run linters only if the modified files are python

* rm main.yaml in CI. Let every config manage itself

Otherwise, cannot restrict linting to just python files.

* Run unit tests concurrently in the CI

* Only lint using `black`. Nothing more.