Refactor logging-related code (#240)
* Replace `logging` with `loguru`

And make most of the logger calls use the same logger instance,
configured in the `log.py` file.

* Use envvar DEODE_LOGLEVEL for loglevel. Formatting

With a fallback to the value of the LOGURU_LEVEL envvar, and a default
to "INFO" if none of these envvars are set.

Some formatting for log messages.

* rm `get_logger_from_config` and `self.logger`

* Make deode executable as `python -m path/to/deode`

* Enable/disable logger according to package's use

Enable if the project is being used as an application.
Disable if the project is being used as a library (leave it for the user
to enable it if then so wish).

This is considered best practice, se, for instance,
<https://loguru.readthedocs.io/en/stable/resources/recipes.html#configuring-loguru-to-be-used-by-a-library-or-an-application>.

* Enable logger in files that use deode as library

* Remove `--loglevel` command line arg

The loglevel should now be especified using the `DEODE_LOGLEVEL`
envvar. For instance,

`DEODE_LOGLEVEL='DEBUG' deode show config`

* Remove opt to set loglevel via config file

One should now use the DEODE_LOGLEVEL envvar for this. E.g.,

`DEODE_LOGLEVEL='DEBUG' deode show config`

* Refctor logs.py

* A little refactoring in log code

* Linting fixes.

* Fix mocking leak bug in smoke test

That would cause some objects to be mocked in every test, while they
should only have been mocked in the module where the mocking was added.

* More descriptive error when task not implemented

* Add `loglevel` back to known config options

* Add `deode_setup` method to `logger`

So we can reconfigure the logger dynamically in an easier fashion.

* Allow tasks to inherit and reconfigure logger opts

* Update pytest-cov to get rid of warnings

* Adapt some new files to  new logger style

* Convert some `deode` imports to relative

* Better (loguru-native) way to configure logger

* Update version of loguru

* Also save logs to `~/.logs/deode` by default

* Set different venv keys for different workflows

The workflows do not necessarily install the same set of dependencies.

* Propagate `loglevel` in `SuiteDefinition`

* Fixes to setting loglevel around the package

* Fix uses of `--config-file` in README's examples