Rewrite `deode doc config` command (#315)
* rm "allOf" from included json schema refs

* Rewrite config_doc.py

Do not include data from `config`, only from the schema.

It is strange documenting things that are not in the schema. This leads to
us not feeling responsible for updating the schema when new properties are
added. It can even lead to misleading results such as when reporting the
`vertical_levels` values from the config as "default", when there is actually no
default in the schema.

The whole point of having the schema is to have a way to define and validate
recognized fields, provide descriptions, define default values, and add general metadata.
If we mix schema and config in the documentation we will end up in a config mess pretty
quickly. Not to mention the nightmare of picking information from different sources
with differentformats without any guaranteed standard.

If we need to check actual values used in the config, we can run `deode show config`.

* Defer doc of json schema do a library

* Linting

* Some formatting changes in the md config doc

* Expand schema refs before generating config doc